Revert type size (requires further investigation), add export w/ decompress feature

+++ type.h
@@ -2,12 +2,12 @@
 #define TYPE_H
 
 typedef unsigned char u8;
-typedef unsigned short int u16;
-typedef unsigned long u32;
-typedef unsigned long long int u64;
+typedef unsigned short u16;
+typedef unsigned int u32;
+typedef unsigned long u64;
 typedef signed char s8;
-typedef signed short int s16;
-typedef signed long s32;
-typedef signed long long int s64;
+typedef signed short s16;
+typedef signed int s32;
+typedef signed long s64;
 
 #endif
